Revert "feat: TypeScript対応"
This reverts commit b7a86a9e79261ccc532179eaabee0273203c1b84.
