+++ basic/server/package.json
@@ -11,10 +11,5 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "express": "^4.17.1"
-  },
-  "devDependencies": {
-    "@types/body-parser": "^1.19.0",
-    "@types/express": "^4.17.5",
-    "typescript": "^3.8.3"
   }
 }
